[DCS-311] Fix cast exception error.

+++ srv/src/main/java/net/ogify/rest/resources/OrderResource.java
@@ -147,7 +147,7 @@ public class OrderResource {
      */
     @GET
     @Path("/{id}/rate")
-    public Long getOrderRating(@PathParam("id") Long orderId) {
+    public Integer getOrderRating(@PathParam("id") Long orderId) {
         return orderProcessor.getUserRateForOrder(userId, orderId);
     }
 }
